+++ crates/algorithms/ecdsa/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#1585](https://github.com/celabshq/libcrux/pull/1585): `p256::Signature::to_der`/`from_der` to encode/decode signatures as DER
+  `ECDSA-Sig-Value`
 - [#1584](https://github.com/celabshq/libcrux/pull/1584): `PrivateKey::public_key` to derive the public key from a P-256 private key
 - [#1584](https://github.com/celabshq/libcrux/pull/1584): `p256::rand::generate_key_pair` to generate a full P-256 key pair